4 -- restore ResolveRecorder but more efficiently
Not removing things from resolving causes the map to eventually grow
to the size of the number of objects, which uses more memory (some
files have millions of indirect objects) and is slower as the map gets
larger. Make ResolveRecorder more efficient by saving the iterator and
using that to erase the item from the set rather than doing a lookup
on the object.